requirements.txt
# requirements.txt (Final Version with All Libraries)
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
python-socketio==5.11.1
Werkzeug==2.3.7
SQLAlchemy==2.0.43
line-bot-sdk==3.19.1
requests==2.32.3
python-dotenv==1.0.0
click==8.1.7
boto3==1.28.60
gevent>=24.2.1
gevent-websocket==0.10.1
Flask-WTF==1.2.1
WTForms==3.1.2
WTForms-SQLAlchemy==0.3
pytz==2024.1
psycopg2-binary
gunicorn
eventlet
bleach==6.1.0
